[refactor] shell_commands 하위 디렉토리 분리
