+++ src/__init__.py
@@ -1,14 +1,14 @@
-from src.shell_commands.help import HelpShellCommand
-from src.shell_commands.read import ReadShellCommand
-from src.shell_commands.write import WriteShellCommand
-from src.shell_commands.full_read import FullReadShellCommand
-from src.shell_commands.full_write import FullWriteShellCommand
-from src.shell_commands.exit import ExitShellCommand
-from src.shell_commands.erase import EraseShellCommand
-from src.shell_commands.flush import FlushShellCommand
-from src.shell_commands.erase_range import EraseRangeShellCommand
+from src.shell_commands.action.help import HelpShellCommand
+from src.shell_commands.action.read import ReadShellCommand
+from src.shell_commands.action.write import WriteShellCommand
+from src.shell_commands.action.full_read import FullReadShellCommand
+from src.shell_commands.action.full_write import FullWriteShellCommand
+from src.shell_commands.action.exit import ExitShellCommand
+from src.shell_commands.action.erase import EraseShellCommand
+from src.shell_commands.action.flush import FlushShellCommand
+from src.shell_commands.action.erase_range import EraseRangeShellCommand
 
 # 테스트 시나리오
-from src.shell_commands.full_write_and_read_compare import FullWriteAndReadCompareShellCommand
-from src.shell_commands.partial_lba_write import PartialLBAWriteShellCommand
-from src.shell_commands.write_read_aging import WriteReadAgingShellCommand
+from src.shell_commands.script.full_write_and_read_compare import FullWriteAndReadCompareShellCommand
+from src.shell_commands.script.partial_lba_write import PartialLBAWriteShellCommand
+from src.shell_commands.script.write_read_aging import WriteReadAgingShellCommand
