+++ src/shell_commands/action/erase.py
