+++ src/shell_commands/action/erase_range.py
