+++ src/shell_commands/action/full_read.py
