+++ src/shell_commands/action/full_write.py
