+++ src/shell_commands/script/erase_and_write_aging.py
