+++ src/shell_commands/script/full_write_and_read_compare.py
