+++ tests/test_full_read.py
@@ -1,7 +1,7 @@
 import pytest
 
 from src.shell_commands.shell_command_action import InvalidArgumentException
-from src.shell_commands.full_read import FullReadShellCommand
+from src.shell_commands.action.full_read import FullReadShellCommand
 from src.ssd_file_manager import SSDFileManager
