+++ tests/test_full_write.py
@@ -1,7 +1,7 @@
 import pytest
 
 from src.shell_commands.shell_command_action import InvalidArgumentException
-from src.shell_commands.full_write import FullWriteShellCommand
+from src.shell_commands.action.full_write import FullWriteShellCommand
 from src.ssd_file_manager import SSDFileManager
 from src.ssd_driver import SSDDriver
