+++ tests/test_partial_lba_write.py
@@ -2,7 +2,7 @@ import pytest
 from pytest_mock import MockerFixture
 
 from src.shell_commands.shell_command_action import InvalidArgumentException
-from src.shell_commands.partial_lba_write import PartialLBAWriteShellCommand
+from src.shell_commands.script.partial_lba_write import PartialLBAWriteShellCommand
 
 
 def create_test_value():
