+++ tests/test_read.py
@@ -1,7 +1,7 @@
 import pytest
 
 from src.shell_commands.shell_command_action import InvalidArgumentException
-from src.shell_commands.read import ReadShellCommand
+from src.shell_commands.action.read import ReadShellCommand
 from src.ssd_driver import SSDDriver
