+++ tests/test_write.py
@@ -1,7 +1,7 @@
 import pytest
 
 from src.shell_commands.shell_command_action import InvalidArgumentException
-from src.shell_commands.write import WriteShellCommand
+from src.shell_commands.action.write import WriteShellCommand
 from src.ssd_driver import SSDDriver
